feat: plugin system enhancements and fixes
- Add PluginConfigWidget and PluginConfigDialog for plugin configuration UI
- Add PluginSetupWizard for guided plugin setup
- Add getEntry() method to PluginConfigModel for entry access
- Fix compilation errors:
  - PluginCommands.h: Include full PluginManager.h (QPointer requires complete type)
  - PDFViewer.cpp: Fix shouldProcessPage() call signature
  - PluginManagerPage.cpp: Use ElaMessageBar::information()
  - PluginConfigDialog.cpp: Use ElaMessageBar::information()
  - DocumentController.cpp: Use getDocumentFilePath()
  - SearchEngine.cpp: Fix SearchResult member names
- Add example plugins and documentation
- Reorganize test structure

+++ app/delegate/AnnotationRenderDelegate.cpp
@@ -212,8 +212,9 @@ void AnnotationRenderDelegate::renderSquiggly(QPainter* painter,
     path.moveTo(x, y);
     while (x < rect.right()) {
         x += wavelength / 2;
-        y = (x / wavelength) % 2 == 0 ? rect.bottom() - amplitude
-                                      : rect.bottom() + amplitude;
+        y = static_cast<int>(x / wavelength) % 2 == 0
+                ? rect.bottom() - amplitude
+                : rect.bottom() + amplitude;
         path.lineTo(x, y);
     }
